# Optimize integer value dumps